refactor: move advanced host options into separate component and add lazy loading
